Move generic thread management functions into helper file
